Adding a few new changes.

+++ App21/App21/IProduct.cs
@@ -9,7 +9,7 @@ namespace App21
     public interface IProduct
     {
         string Name { get; }
-        abstract float GetPrice(float weightInGrams);
-        abstract float GetPrice();
+        abstract decimal GetPrice(float weightInGrams);
+        abstract decimal GetPrice();
     }
 }
